feat: 댓글 수 기능 구현 및 하드코딩 제거(#111)
- comment list 조회 시 댓글 수 필드 추가
- post list 조회 시 하드코딩된 댓글 수를 실제 DB 조회로 변경

+++ src/main/java/com/develop_ping/union/post/presentation/dto/response/PostListResponse.java
@@ -18,8 +18,8 @@ public class PostListResponse {
     private ZonedDateTime createdAt;
     private AuthorResponse author;
     private Integer views;
-    private Integer postLikes;
-    private Integer commentCount;
+    private long postLikes;
+    private long commentCount;
 
     @Builder
     public PostListResponse(Long id,
@@ -29,8 +29,8 @@ public class PostListResponse {
                             ZonedDateTime createdAt,
                             AuthorResponse author,
                             Integer views,
-                            Integer postLikes,
-                            Integer commentCount) {
+                            long postLikes,
+                            long commentCount) {
         this.id = id;
         this.title = title;
         this.contentPreview = contentPreview;
